Design home screen architecture
Create VIPER template for home screen that include ViewController, Presenter, Interactor, Router, entity and protocol files. ViewController’s main purpose is to display info to the user and detect user interaction.
The presenter receives user action information from the view/view controller and is responsible for the UI related  business logic data
The interactor has knowledge of our models, or entities, and fetches and stores data from a manager class
The entities are our data models (i.e. MapName, MapCoordinate, MapLocation)
The router has one point of contact: the presenter. The router  owns a window and takes care of the navigation in a module